requirements.txt
cffi==1.17.1
pillow==11.0.0
pycparser==2.22
pyheif==0.8.0
colorama==0.4.6
flask==3.1.0
apscheduler==3.11.0
gunicorn==23.0.0